refactor: move package
- add package util, strategy
