+++ src/main/java/racingcar/domain/Car.java
@@ -1,5 +1,7 @@
 package racingcar.domain;
 
+import racingcar.strategy.Generator;
+
 public class Car {
     private static final int MINIMUM_MOVE_VALUE = 4;
