+++ src/main/java/racingcar/strategy/RandomValueGenerator.java
@@ -1,4 +1,4 @@
-package racingcar.domain;
+package racingcar.strategy;
 
 import java.util.Random;
