+++ src/test/java/racingcar/domain/CarTest.java
@@ -4,6 +4,7 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
+import racingcar.strategy.ManualValueGenerator;
 
 import static org.assertj.core.api.Assertions.assertThat;
